New Images + Alexicia's Voldemort/Spell changes
This updates the images for Beater, Seeker, Voldemort, and Spells.
Alexicia also sent me her changes to the Voldemort, Projectiles, and
Spells classes. This also comments out the music because it got annoying
after compiling for the 12th time. This also adds a setMovement method
to SmoothMover that we forgot to push in class. There are definitely
some bugs lurking around the high score board, but I don't understand
how to fix them.